agentic_ai_prototype/requirements.txt
streamlit>=1.24.0
SpeechRecognition>=3.10.0
gTTS>=2.3.2
PyAudio>=0.2.13
python-dotenv>=1.0.0 